yoga_payment_bot/requirements.txt
anyio==3.7.1
APScheduler==3.9.1.post1
asgiref==3.8.1
cachetools==5.0.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==3.2.16
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-telegram-bot==21.3
pytz==2024.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tornado==6.4.1
typing-extensions==4.12.2
tzlocal==5.2
